fix: polyfill-loader
