+++ packages/build-user-config/src/userConfig/polyfill.js
@@ -1,5 +1,6 @@
 const { injectTransformRuntime } = require('@builder/app-helpers');
 const processPresetEnv  = require('../utils/processPresetEnv');
+const addBablePlugins = require('../utils/addBabelPlugins');
 
 module.exports = (config, polyfill) => {
   const presetEnvParam = {
@@ -9,6 +10,30 @@ module.exports = (config, polyfill) => {
     presetEnvParam.corejs = 3;
     if (polyfill === 'usage') {
       injectTransformRuntime(config);
+    } else if (polyfill === 'entry') {
+      const entries = config.toConfig().entry;
+      const rule = config.module.rule('polyfill').test(/\.jsx?|\.tsx?$/);
+      const fileList = [];
+      if (!entries) return;
+      Object.keys(entries).forEach((key) => {
+        let addPolyfill = false;
+        // only include entry path
+        for (let i = 0; i < entries[key].length; i += 1) {
+          // filter node_modules file add by plugin
+          if (!/node_modules/.test(entries[key][i])) {
+            rule.include.add(entries[key][i]);
+            fileList.push(entries[key][i]);
+            addPolyfill = true;
+            break;
+          }
+        }
+        if (!addPolyfill) {
+          rule.include.add(entries[key][0]);
+          fileList.push(entries[key][0]);
+        }
+      });
+      rule.use('polyfill-loader').loader(require.resolve('../utils/polyfillLoader')).options({});
+      addBablePlugins(config, [[require.resolve('../utils/babelPluginCorejsLock.js'), { fileList }]]);
     }
   }
   processPresetEnv(config, presetEnvParam);
