+++ packages/plugin-react-app/src/index.js
@@ -2,7 +2,7 @@ const path = require('path');
 const { applyCliOption, applyUserConfig, getEnhancedWebpackConfig } = require('@builder/user-config');
 const { getWebpackConfig, getBabelConfig } = require('build-scripts-config');
 const { WEB, MINIAPP, WECHAT_MINIPROGRAM} = require('./constants');
-const customConfigs = require('./config');
+const getCustomConfigs = require('./config');
 const setBase = require('./setBase');
 const setDev = require('./setDev');
 const setBuild = require('./setBuild');
@@ -25,7 +25,7 @@ module.exports = (api) => {
   applyCliOption(api);
 
   // register user config
-  applyUserConfig(api, { customConfigs });
+  applyUserConfig(api, { customConfigs: getCustomConfigs(userConfig) });
 
   // set webpack config
   onGetWebpackConfig(chainConfig => {
